requirements.txt
websockets
asyncio
python-dotenv
kafka-python
pandas
polars
apache-airflow
apache-airflow-providers-cncf-kubernetes
apache-airflow-providers-docker
pytest
black
flake8
pydantic>=2.0
mypy
